\name{Open_library}
\alias{Open_library}
\title{Open_library()}
\usage{
Open_library()
}
\description{
Look for each .txt file inside of a directory and open each one to create a long format dataframe containing each spectra. Made for .txt files created by the ViewSpec Pro software. It can open both files with or without metadata.
}
\examples{
Open_library(df)
}
