\name{Open_library}
\alias{Open_library}
\title{Open_library()}
\usage{
Open_library()
}
\description{
Search for each .txt file within a directory and open them to create a long-format DataFrame containing the spectra data. Designed for .txt files generated by the ViewSpec Pro software, this process can handle files both with and without metadata.
}
\examples{
Open_library(df)
}